Its a FLoat!

+++ src/main.cpp
@@ -9,7 +9,7 @@ AH_EasyDriver stepperSlider(200,32,33,27,26,28);
 int direction = 1;
 int rotationSlider = 5;
 int rotationCamera = 1;
-int afterTurn = 1;
+float afterTurn = 1;
 
 void setup(){
   //Steppermotorsetup
@@ -34,7 +34,7 @@ void loop(){
     if(afterTurn < 1){
       afterTurn = afterTurn + 0.1;
     }
-    
+
     stepperCamera.setSpeedRPM(100 * afterTurn);
     stepperSlider.setSpeedRPM(100 * afterTurn);
     stepperCamera.rotate(direction * rotationCamera);
